move out of single class
