+++ starsky/starskytest/starsky.foundation.database/QueryTest/QueryTest.cs
@@ -166,87 +166,6 @@ namespace starskytest.starsky.foundation.database.QueryTest
 			}
         }
 
-        [TestMethod]
-        public void QueryAddSingleItemRootFolderTest()
-        {
-
-            InsertSearchData();
-            // Test root folder ("/)
-            var getAllFilesExpectedResult = new List<FileIndexItem>
-            {
-                _insertSearchDatahiJpgInput,
-                _insertSearchDatahi2JpgInput,
-                _insertSearchDatahi3JpgInput,
-                _insertSearchDatahi4JpgInput
-            };
-
-            var getAllResult = _query.GetAllFiles("/basic");
-
-            CollectionAssert.AreEqual(getAllFilesExpectedResult.Select(p => p.FilePath).ToList(), 
-                getAllResult.Select(p => p.FilePath).ToList());
-        }
-
-        [TestMethod]
-        public void GetAllFiles_DisposedItem()
-        {
-	        var serviceScope = CreateNewScope();
-	        var scope = serviceScope.CreateScope();
-	        var dbContext = scope.ServiceProvider.GetRequiredService<ApplicationDbContext>();
-	        var query = new global::starsky.foundation.database.Query.Query(dbContext,_memoryCache, new AppSettings(), serviceScope);
-	        
-	        // item sub folder
-	        var item = new FileIndexItem("/test_821827/test_0191919.jpg");
-	        dbContext.FileIndex.Add(item);
-	        dbContext.SaveChanges();
-	        
-	        // Important to dispose!
-	        dbContext.Dispose();
-
-	        item.Tags = "test";
-	        query.UpdateItem(item);
-
-	        var getItem = query.GetAllFiles("/test_821827");
-	        Assert.IsNotNull(getItem);
-	        Assert.AreEqual("test", getItem.FirstOrDefault().Tags);
-
-	        query.RemoveItem(getItem.FirstOrDefault());
-        }
-
-        [TestMethod]
-        public void QueryAddSingleItemSubFolderTest()
-        {
-            InsertSearchData();
-
-            // Test subfolder
-            var getAllFilesSubFolderExpectedResult = new List<FileIndexItem> {_insertSearchDatahi2SubfolderJpgInput};
-
-            var getAllResultSubfolder = _query.GetAllFiles("/basic/subfolder");
-            
-            CollectionAssert.AreEqual(getAllFilesSubFolderExpectedResult.Select(p => p.FilePath).ToList(), 
-                getAllResultSubfolder.Select(p => p.FilePath).ToList());
-        }
-
-        [TestMethod]
-        public async Task GetAllFilesAsync_GetResult()
-        {
-	        var appSettings = new AppSettings
-	        {
-		        DatabaseType = AppSettings.DatabaseTypeList.InMemoryDatabase
-	        };
-	        var dbContext = new SetupDatabaseTypes(appSettings, null).BuilderDbFactory();
-	        var query = new Query(dbContext);
-
-	        await dbContext.FileIndex.AddAsync(new FileIndexItem("/GetAllFilesAsync") {IsDirectory = true});
-	        await dbContext.FileIndex.AddAsync(new FileIndexItem("/GetAllFilesAsync/test") {IsDirectory = true});
-	        await dbContext.FileIndex.AddAsync(new FileIndexItem("/GetAllFilesAsync/test.jpg"));
-	        await dbContext.FileIndex.AddAsync(new FileIndexItem("/GetAllFilesAsync/test/test.jpg"));
-	        await dbContext.SaveChangesAsync();
-	        
-	        var items = await query.GetAllFilesAsync("/GetAllFilesAsync");
-
-	        Assert.AreEqual("/GetAllFilesAsync/test.jpg", items[0].FilePath);
-	        Assert.AreEqual(FileIndexItem.ExifStatus.Ok, items[0].Status);
-        }
         
         [TestMethod]
         public async Task GetAllFilesAsync_Disposed()
